arcflexSFDC 2_0_1_2922:
This build fixes the issue where users were unable to log on to their next works order until their last one had finished posting. Start times from the tempoary table are added to the production report loading by the client, and can therefore now be deleted from the tempoary table prior to posting, rather than at the end of the loading. In the process we have reduced the size of the Priority loading from 18 steps to 5.

Additionally I have included some of the modifications requested to this module by GSM, namely that the report of a "Reject" quantity will now spawn a new Reject line. This enables the user to reject multiple items from the same batch with different reason codes. To acheive this there is a slight change in functionality of the load - Priority will now show one production entry for the labour start and end, and a seperate line for each quantity (approve/mrb/reject).

I have also implimented the parallel works order processing that was requested by Arcflex. When a user begins a report they will still get the "logged on to job XYZ" message, but this is now a yes/no dialog. If they choose not to open the suggested job they can simply proceed and log onto another new job. However the device will prompt yes/no for each job the user is currently logged into. So the next time they open a production report they will be prompted to re-open each of their previously opened works orders, before the device will let them create a new one. Hopefully this will give the flexibility required while still allowing users to keep track of which jobs they have placed a start time marker against.